_ARCHIVED/projects/Frontline AI/Frontline AI/atlas-mvp/backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
websockets>=12.0
opencv-python
numpy
pyyaml
sqlalchemy>=2.0.23
aiosqlite>=0.19.0
python-multipart>=0.0.6
pillow>=10.0.1
pydantic>=2.4.2

# Audio processing dependencies
pyaudio>=0.2.11
librosa>=0.10.0
torch>=2.0.0
torchaudio>=2.0.0
torchvision>=0.15.0

# Vision and optimization dependencies
ultralytics>=8.0.0
coremltools>=6.3
